Delete excess file and add text and functions to file
There was a mistype on one of the initial recipe types which had put the text on two documents. This had led to the deletion of one. Multipel additions had been made to the other file. These additions include, the correction of the image which ahd not displayed due to wrong filetype and its location being set to ../odin-recipes instead of ../images. Other additions that were made include the creation of directions and ingredients lists with coorosponding headings.